+++ CHANGES.md
@@ -13,6 +13,7 @@ Bugfix 🐛:
  - Fix a problem with database migration on nightly builds (#3335)
  - Implement a workaround to render &lt;del&gt; and &lt;u&gt; in the timeline (#1817)
  - Make sure the SDK can retrieve the secret storage if the system is upgraded (#3304)
+ - Spaces | Explore room list: the RoomId is displayed instead of name (#3371)
 
 Translations 🗣:
  -